Consolidate code into single script
