--- src/default_data.js
@@ -1,4 +1,0 @@
-const DEFAULT_DATA = {
-  lists: DEFAULT_LIST,
-  state: {}
-};
