+++ src/index.html
@@ -330,13 +330,6 @@
   <!-- External script loading order is important -->
   <!-- Load list data files first -->
   <script src="default_list.js"></script>
-  <script src="default_data.js"></script>
-  <!-- Load submodules before main script -->
-  <script src="lib/promptUtils.js"></script>
-  <script src="listManager.js"></script>
-  <script src="stateManager.js"></script>
-  <script src="storageManager.js"></script>
-  <script src="uiControls.js"></script>
   <!-- Then load main application logic -->
   <script src="script.js"></script>
 </body>
