--- src/lib/AGENTS.md
@@ -1,4 +1,0 @@
-# AGENT Instructions for lib
-Modules in this directory must remain side-effect free. Do not access the DOM or mutate global state during import. Export pure functions only using CommonJS `module.exports`.
-
-All code here is guaranteed to be pure functional programming. Keep functions well organized and name them expressively so their purpose is obvious.
